requirements.txt
argparse==1.2.1
backports.ssl-match-hostname==3.4.0.2
gevent==1.0.1
gevent-socketio==0.3.6
gevent-websocket==0.9.3
greenlet==0.4.2
requests==2.3.0
six==1.7.2
socketIO-client==0.5.3
websocket-client==0.15.0
wsgiref==0.1.2
